@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.10.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,40 +30,41 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^6.0.5
   flutter_screenutil: ^5.8.4
-  google_fonts: ^5.1.0
-  flutter_animate: ^4.2.0+1
-  image_picker: ^1.0.2
-  image_cropper: ^5.0.0
+  google_fonts: ^6.1.0
+  flutter_animate: ^4.5.0
+  image_picker: ^1.0.7
+  image_cropper: ^5.0.1
   smooth_page_indicator: ^1.1.0
-  shared_preferences: ^2.2.0
+  shared_preferences: ^2.2.2
   ionicons: ^0.2.2
-  flutter_zoom_drawer: ^3.1.1
+  flutter_zoom_drawer: ^3.1.2
   dartz: ^0.10.1
   flutter_chat_bubble: ^2.0.2
-  flutter_svg: ^2.0.7
+  flutter_svg: ^2.0.9
   feather_icons: ^1.2.0
-  cached_network_image: ^3.2.3
+  cached_network_image: ^3.3.1
   equatable: ^2.0.5
-  intl_phone_field: ^3.2.0
-  country_picker: ^2.0.20
-  firebase_core: ^2.15.0
-  firebase_storage: ^11.2.5
+  country_picker: ^2.0.24
+  firebase_core: ^2.25.4
+  firebase_storage: ^11.6.5
   socket_io_client: ^2.0.3+1
-  flutter_slidable: ^3.0.0
+  flutter_slidable: ^3.0.1
   bloc: ^8.1.2
-  meta: ^1.9.1
-  go_router: ^10.0.0
+  go_router: ^13.2.0
+  url_launcher: ^6.2.4
   flutter_bloc: ^8.1.3
-  dio: ^5.3.2
-  get_it: ^7.6.0
-  uuid: ^3.0.7
-  intl: ^0.18.1
+  dio: ^5.4.0
+  get_it: ^7.6.7
+  uuid: ^4.3.3
+  intl: ^0.19.0
+  meta: ^1.10.0
+  file_picker: ^6.1.1
+  intl_phone_field: ^3.2.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.5
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
@@ -74,7 +75,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.2
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
