@@ -2,21 +2,21 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:job_hub/core/utils/constants.dart';
-import 'package:job_hub/core/utils/routes_config/app_routes.dart';
-import 'package:job_hub/core/widgets/text_styles/app_style.dart';
-import 'package:job_hub/core/utils/cache_helper.dart';
-import 'package:job_hub/core/widgets/spacers/height_spacer.dart';
-import 'package:job_hub/core/widgets/text_styles/reusable_text.dart';
-import 'package:job_hub/core/widgets/text_fields/email_text_field.dart';
+import 'package:job_hub/core/functions/handle_cache.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:job_hub/core/functions/show_snack_bar.dart';
+import 'package:job_hub/core/widgets/text_styles/app_style.dart';
+import 'package:job_hub/core/widgets/spacers/height_spacer.dart';
+import 'package:job_hub/core/utils/routes_config/app_routes.dart';
+import 'package:job_hub/core/widgets/text_styles/reusable_text.dart';
+import 'package:job_hub/core/widgets/text_fields/email_text_field.dart';
 import 'package:job_hub/core/widgets/text_fields/password_text_field.dart';
 import 'package:job_hub/features/auth/presentation/cubits/login_cubit/login_cubit.dart';
 import 'package:job_hub/features/auth/presentation/views/login_view/widgets/login_button.dart';
 import 'package:job_hub/features/auth/presentation/views/login_view/widgets/create_account_text.dart';
 
 class LoginViewBody extends StatelessWidget {
-  const LoginViewBody({Key? key}) : super(key: key);
+  const LoginViewBody({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -24,18 +24,9 @@ class LoginViewBody extends StatelessWidget {
     return BlocConsumer<LoginCubit, LoginState>(
       listener: (context, state) async {
         if (state is LoginSuccess) {
-          showCustomSnackBar(context, 'Login Successful');
-          token = state.response.token;
-          userId = state.response.userData!.id!;
-          userImage = state.response.userData!.imageUrl!;
           GoRouter.of(context).pushReplacement(AppRoutes.kMainView);
-          await Future.wait([
-            CacheHelper.setData(key: 'token', value: state.response.token),
-            CacheHelper.setData(
-                key: 'userId', value: state.response.userData!.id),
-            CacheHelper.setData(
-                key: 'userImage', value: state.response.userData!.imageUrl)
-          ]);
+          showCustomSnackBar(context, 'Login Successful');
+          await setCache(state.response);
         } else if (state is LoginFailure) {
           showCustomSnackBar(context, 'Error: ${state.errMessage}');
         }
@@ -53,7 +44,7 @@ class LoginViewBody extends StatelessWidget {
                 ReusableText(
                   text: 'Welcome Back!',
                   style: appStyle(
-                    30,
+                    24,
                     Color(kDark.value),
                     FontWeight.w600,
                   ),
@@ -64,7 +55,7 @@ class LoginViewBody extends StatelessWidget {
                   style: appStyle(
                     16,
                     Color(kDarkGrey.value),
-                    FontWeight.w600,
+                    FontWeight.w500,
                   ),
                 ),
                 const HeightSpacer(size: 40),
@@ -76,7 +67,7 @@ class LoginViewBody extends StatelessWidget {
                   isObscure: cubit.isObscure,
                 ),
                 const HeightSpacer(size: 20),
-                const CreateAccountText(),
+                CreateAccountText(cubit: cubit),
                 const HeightSpacer(size: 20),
                 LoginButton(state: state, cubit: cubit),
               ],
