@@ -1,27 +1,33 @@
 import 'dart:convert';
-
 import 'package:equatable/equatable.dart';
 
 class SendMessageRequest extends Equatable {
   final String content;
-  final String receiver;
-  final String chat;
+  final String receiverId;
+  final String chatId;
+  final String? message;
 
-  const SendMessageRequest(
-      {required this.content, required this.receiver, required this.chat});
+  const SendMessageRequest({
+    required this.content,
+    required this.receiverId,
+    required this.chatId,
+    this.message,
+  });
 
   factory SendMessageRequest.fromMap(Map<String, dynamic> data) {
     return SendMessageRequest(
       content: data['content'] as String,
-      receiver: data['receiver'] as String,
-      chat: data['chat'] as String,
+      receiverId: data['receiverId'] as String,
+      chatId: data['chatId'] as String,
+      message: data['message'] as String?,
     );
   }
 
   Map<String, dynamic> toMap() => {
         'content': content,
-        'receiver': receiver,
-        'chat': chat,
+        'receiverId': receiverId,
+        'chatId': chatId,
+        'message': message,
       };
 
   factory SendMessageRequest.fromJson(String data) {
@@ -32,5 +38,5 @@ class SendMessageRequest extends Equatable {
   String toJson() => json.encode(toMap());
 
   @override
-  List<Object?> get props => [content, receiver, chat];
+  List<Object?> get props => [content, receiverId, chatId];
 }
