@@ -1,31 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:job_hub/core/utils/constants.dart';
-import 'package:job_hub/core/widgets/custom_app_bar.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:job_hub/features/main/presentation/views/widgets/leading_button.dart';
+import 'package:job_hub/features/drawer/presentation/views/widgets/leading_button.dart';
 
-class HomeAppBar extends StatelessWidget {
-  const HomeAppBar({Key? key}) : super(key: key);
+class HomeAppBar extends StatelessWidget implements PreferredSizeWidget {
+  const HomeAppBar({super.key});
+
+  @override
+  Size get preferredSize => const Size.fromHeight(50);
 
   @override
   Widget build(BuildContext context) {
-    return CustomAppBar(
+    return AppBar(
+      elevation: 0.0,
+      scrolledUnderElevation: 0.0,
       actions: [
         Padding(
-          padding: EdgeInsets.all(
-            12.h,
-          ),
+          padding: EdgeInsets.all(12.h),
           child: CircleAvatar(
             radius: 15,
             backgroundImage: CachedNetworkImageProvider(userImage!),
           ),
         ),
       ],
-      leading: Padding(
-        padding: EdgeInsets.all(12.h),
-        child: const LeadingButton(),
-      ),
+      leading: const LeadingButton(),
     );
   }
 }
