@@ -8,7 +8,6 @@ class ApiService {
   Future<Map<String, dynamic>> post({
     required String endpoint,
     Map<String, dynamic>? data,
-    Map<String, dynamic>? query,
     String? token,
   }) async {
     _dio.options.headers = {
@@ -18,24 +17,18 @@ class ApiService {
     var response = await _dio.post(
       endpoint,
       data: data,
-      queryParameters: query ?? {},
     );
     return response.data;
   }
 
   Future<Map<String, dynamic>> get({
     required String endpoint,
-    Map<String, dynamic>? data,
     Map<String, dynamic>? query,
     String? token,
   }) async {
-    _dio.options.headers = {
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer $token'
-    };
+    _dio.options.headers = {'Authorization': 'Bearer $token'};
     var response = await _dio.get(
       endpoint,
-      data: data,
       queryParameters: query ?? {},
     );
     return response.data;
@@ -51,6 +44,22 @@ class ApiService {
       'Authorization': 'Bearer $token'
     };
     var response = await _dio.delete(
+      endpoint,
+      data: data,
+    );
+    return response.data;
+  }
+
+  Future<Map<String, dynamic>> put({
+    required String endpoint,
+    Map<String, dynamic>? data,
+    String? token,
+  }) async {
+    _dio.options.headers = {
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer $token'
+    };
+    var response = await _dio.put(
       endpoint,
       data: data,
     );
