@@ -1,6 +1,7 @@
-import '../text_styles/app_style.dart';
 import 'package:flutter/material.dart';
 import 'package:job_hub/core/utils/constants.dart';
+import 'package:job_hub/core/widgets/text_styles/app_style.dart';
+import 'package:job_hub/core/widgets/text_styles/reusable_text.dart';
 
 class CustomErrorWidget extends StatelessWidget {
   const CustomErrorWidget({super.key, required this.message});
@@ -10,10 +11,10 @@ class CustomErrorWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Text(
-        message,
+      child: ReusableText(
+        text: message,
         style: appStyle(
-          16,
+          18,
           Color(kDark.value),
           FontWeight.w500,
         ),
