@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:job_hub/core/utils/constants.dart';
 import 'package:job_hub/core/widgets/buttons/custom_button.dart';
 import 'package:job_hub/features/auth/data/models/login_models/login_request.dart';
 import 'package:job_hub/features/auth/presentation/cubits/login_cubit/login_cubit.dart';
@@ -16,7 +17,7 @@ class LoginButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return state is LoginLoading
-        ? const Center(child: CircularProgressIndicator())
+        ? Center(child: CircularProgressIndicator(color: Color(kOrange.value)))
         : CustomButton(
             text: 'Login',
             onTap: () async {
