@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:job_hub/features/bookmarks/presentation/views/bookmarks_view/widgets/bookmarks_app_bar.dart';
 import 'package:job_hub/features/bookmarks/presentation/views/bookmarks_view/widgets/bookmarks_view_body.dart';
 
@@ -8,12 +7,12 @@ class BookmarksView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(50.h),
-        child: const BookmarksAppBar(),
+    return const SafeArea(
+      top: false,
+      child: Scaffold(
+        appBar: BookmarksAppBar(),
+        body: BookmarksViewBody(),
       ),
-      body: const BookmarksViewBody(),
     );
   }
 }
