@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'package:equatable/equatable.dart';
-import 'package:job_hub/features/auth/data/models/update_user_models/user_data.dart';
 import 'package:job_hub/features/chat/data/models/chat_models/chat.dart';
+import 'package:job_hub/features/profile/data/models/update_user_models/user_data.dart';
 
 class Message extends Equatable {
   final String? id;
@@ -9,10 +9,7 @@ class Message extends Equatable {
   final String? content;
   final UserData? receiver;
   final Chat? chat;
-  final List<dynamic>? readBy;
   final DateTime? createdAt;
-  final DateTime? updatedAt;
-  final int? v;
 
   const Message({
     this.id,
@@ -20,10 +17,7 @@ class Message extends Equatable {
     this.content,
     this.receiver,
     this.chat,
-    this.readBy,
     this.createdAt,
-    this.updatedAt,
-    this.v,
   });
 
   factory Message.fromMap(Map<String, dynamic> data) => Message(
@@ -38,14 +32,9 @@ class Message extends Equatable {
         chat: data['chat'] == null
             ? null
             : Chat.fromMap(data['chat'] as Map<String, dynamic>),
-        readBy: data['readBy'] as List<dynamic>?,
         createdAt: data['createdAt'] == null
             ? null
             : DateTime.parse(data['createdAt'] as String),
-        updatedAt: data['updatedAt'] == null
-            ? null
-            : DateTime.parse(data['updatedAt'] as String),
-        v: data['__v'] as int?,
       );
 
   Map<String, dynamic> toMap() => {
@@ -54,10 +43,7 @@ class Message extends Equatable {
         'content': content,
         'receiver': receiver,
         'chat': chat,
-        'readBy': readBy,
         'createdAt': createdAt?.toIso8601String(),
-        'updatedAt': updatedAt?.toIso8601String(),
-        '__v': v,
       };
 
   factory Message.fromJson(String data) {
@@ -74,10 +60,7 @@ class Message extends Equatable {
       content,
       receiver,
       chat,
-      readBy,
       createdAt,
-      updatedAt,
-      v,
     ];
   }
 }
