@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:job_hub/features/chat/presentation/views/chat_view/widgets/chat_view_app_bar.dart';
 import 'package:job_hub/features/chat/presentation/views/chat_view/widgets/chat_view_body.dart';
 
@@ -8,12 +7,12 @@ class ChatView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(50.h),
-        child: const ChatViewAppBar(),
+    return const SafeArea(
+      top: false,
+      child: Scaffold(
+        appBar: ChatViewAppBar(),
+        body: ChatViewBody(),
       ),
-      body: const ChatViewBody(),
     );
   }
 }
