@@ -1,32 +1,39 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:job_hub/core/utils/constants.dart';
-import 'package:job_hub/core/widgets/custom_app_bar.dart';
-import 'package:job_hub/features/main/presentation/views/widgets/leading_button.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:job_hub/core/widgets/text_styles/app_style.dart';
+import 'package:job_hub/features/drawer/presentation/views/widgets/leading_button.dart';
 
-class ChatViewAppBar extends StatelessWidget {
+class ChatViewAppBar extends StatelessWidget implements PreferredSizeWidget {
   const ChatViewAppBar({super.key});
 
   @override
+  Size get preferredSize => const Size.fromHeight(50);
+
+  @override
   Widget build(BuildContext context) {
-    return CustomAppBar(
-      title: 'Chats',
+    return AppBar(
+      centerTitle: true,
+      scrolledUnderElevation: 0.0,
+      title: Text(
+        'Chats',
+        style: appStyle(
+          16,
+          Color(kDark.value),
+          FontWeight.w600,
+        ),
+      ),
       actions: [
         Padding(
-          padding: EdgeInsets.all(
-            12.h,
-          ),
+          padding: EdgeInsets.all(12.h),
           child: CircleAvatar(
             radius: 15,
             backgroundImage: CachedNetworkImageProvider(userImage!),
           ),
         ),
       ],
-      leading: Padding(
-        padding: EdgeInsets.all(12.h),
-        child: const LeadingButton(),
-      ),
+      leading: const LeadingButton(),
     );
   }
 }
