@@ -7,25 +7,19 @@ class JobsInitial extends JobsState {}
 
 class GetAllJobsLoading extends JobsState {}
 
-class GetAllJobsFailure extends JobsState {
-  final String errMessage;
-
-  GetAllJobsFailure({required this.errMessage});
-}
-
 class GetAllJobsSuccess extends JobsState {
   final List<Job> jobs;
 
   GetAllJobsSuccess({required this.jobs});
 }
 
-class GetJobLoading extends JobsState {}
-
-class GetJobFailure extends JobsState {
+class GetAllJobsFailure extends JobsState {
   final String errMessage;
 
-  GetJobFailure({required this.errMessage});
+  GetAllJobsFailure({required this.errMessage});
 }
+
+class GetJobLoading extends JobsState {}
 
 class GetJobSuccess extends JobsState {
   final Job job;
@@ -33,16 +27,32 @@ class GetJobSuccess extends JobsState {
   GetJobSuccess({required this.job});
 }
 
-class SearchJobsLoading extends JobsState {}
-
-class SearchJobsFailure extends JobsState {
+class GetJobFailure extends JobsState {
   final String errMessage;
 
-  SearchJobsFailure({required this.errMessage});
+  GetJobFailure({required this.errMessage});
 }
+
+class DeleteJobLoading extends JobsState {}
+
+class DeleteJobSuccess extends JobsState {}
+
+class DeleteJobFailure extends JobsState {
+  final String errMessage;
+
+  DeleteJobFailure({required this.errMessage});
+}
+
+class SearchJobsLoading extends JobsState {}
 
 class SearchJobsSuccess extends JobsState {
   final List<dynamic> searchResults;
 
   SearchJobsSuccess({required this.searchResults});
+}
+
+class SearchJobsFailure extends JobsState {
+  final String errMessage;
+
+  SearchJobsFailure({required this.errMessage});
 }
