@@ -4,6 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 import 'package:job_hub/core/utils/constants.dart';
 import 'package:job_hub/core/utils/routes_config/app_routes.dart';
+import 'package:job_hub/core/widgets/spacers/height_spacer.dart';
 import 'package:job_hub/core/widgets/text_styles/app_style.dart';
 import 'package:job_hub/core/widgets/text_styles/reusable_text.dart';
 import 'package:job_hub/core/widgets/spacers/width_spacer.dart';
@@ -16,86 +17,73 @@ class JobTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(bottom: 12.h),
-      child: GestureDetector(
-        onTap: () {
-          GoRouter.of(context).push(AppRoutes.kJobDetailsView, extra: job);
-        },
-        child: Container(
-          padding: EdgeInsets.symmetric(vertical: 10.h, horizontal: 12.w),
-          height: height * 0.15,
-          width: width,
+    return GestureDetector(
+      onTap: () {
+        GoRouter.of(context).push(AppRoutes.kJobDetailsView, extra: job);
+      },
+      child: Container(
+        padding: EdgeInsets.symmetric(horizontal: 8.w, vertical: 12.h),
+        decoration: BoxDecoration(
           color: Color(kLightGrey.value),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  CircleAvatar(
-                    radius: 30,
-                    backgroundImage: CachedNetworkImageProvider(job.imageUrl!),
-                  ),
-                  const WidthSpacer(width: 10),
-                  Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      ReusableText(
-                        text: job.company!,
-                        style: appStyle(
-                          20,
-                          Color(kDark.value),
-                          FontWeight.w600,
-                        ),
-                      ),
-                      SizedBox(
-                        width: width * 0.5,
-                        child: Text(
-                          job.title!,
-                          style: appStyle(
-                            20,
-                            Color(kDark.value),
-                            FontWeight.w600,
-                          ),
-                        ),
-                      )
-                    ],
-                  ),
-                  CircleAvatar(
-                    radius: 18,
-                    child: Icon(
-                      Icons.arrow_forward_ios_outlined,
-                      color: Color(kOrange.value),
-                    ),
-                  )
-                ],
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
+          borderRadius: BorderRadius.circular(12.r),
+        ),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            CircleAvatar(
+              radius: 30.r,
+              backgroundImage: CachedNetworkImageProvider(job.imageUrl!),
+            ),
+            const WidthSpacer(width: 10),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   ReusableText(
-                    text: job.salary!,
-                    style: appStyle(
-                      22,
-                      Color(kDark.value),
-                      FontWeight.w600,
-                    ),
-                  ),
-                  ReusableText(
-                    text: '/${job.contract!}',
+                    text: job.title!,
                     style: appStyle(
                       20,
-                      Color(kDarkGrey.value),
+                      Color(kDark.value),
                       FontWeight.w600,
-                    ),
+                    ).copyWith(overflow: TextOverflow.ellipsis),
+                  ),
+                  ReusableText(
+                    text: job.company!,
+                    style: appStyle(
+                      18,
+                      Color(kDark.value),
+                      FontWeight.w500,
+                    ).copyWith(overflow: TextOverflow.ellipsis),
+                  ),
+                  const HeightSpacer(size: 8),
+                  Row(
+                    children: [
+                      SizedBox(
+                        width: width * 0.6,
+                        child: ReusableText(
+                          text: job.salary!,
+                          style: appStyle(
+                            18,
+                            Color(kDark.value),
+                            FontWeight.w400,
+                          ),
+                        ),
+                      ),
+                    ],
                   ),
                 ],
               ),
-            ],
-          ),
+            ),
+            const WidthSpacer(width: 2),
+            CircleAvatar(
+              radius: 20.r,
+              backgroundColor: Color(kOrange.value),
+              child: const Icon(
+                Icons.arrow_forward_ios_outlined,
+                color: Colors.white,
+              ),
+            ),
+          ],
         ),
       ),
     );
