@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:job_hub/features/home/data/models/job.dart';
 import 'package:job_hub/features/home/presentation/views/job_details_view/widgets/job_details_view_app_bar.dart';
 import 'package:job_hub/features/home/presentation/views/job_details_view/widgets/job_details_view_body.dart';
@@ -12,9 +11,9 @@ class JobDetailsView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(50.h),
-        child: JobsDetailsViewAppBar(company: job.company!),
+      appBar: JobsDetailsViewAppBar(
+        company: job.company!,
+        jobId: job.id!,
       ),
       body: JobDetailsViewBody(job: job),
     );
