@@ -1,21 +1,24 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 import 'package:job_hub/core/widgets/custom_app_bar.dart';
+import 'package:job_hub/features/home/presentation/views/job_details_view/widgets/add_bookmark_button.dart';
 
-class JobsDetailsViewAppBar extends StatelessWidget {
-  const JobsDetailsViewAppBar({super.key, required this.company});
+class JobsDetailsViewAppBar extends StatelessWidget
+    implements PreferredSizeWidget {
+  const JobsDetailsViewAppBar({
+    super.key,
+    required this.company,
+    required this.jobId,
+  });
 
   final String company;
+  final String jobId;
 
   @override
   Widget build(BuildContext context) {
     return CustomAppBar(
-      actions: const [
-        Padding(
-          padding: EdgeInsets.only(right: 8.0),
-          child: Icon(CupertinoIcons.bookmark_solid),
-        )
-      ],
+      actions: [AddBookmarkButton(jobId: jobId)],
       leading: GestureDetector(
         onTap: () {
           GoRouter.of(context).pop();
@@ -25,4 +28,7 @@ class JobsDetailsViewAppBar extends StatelessWidget {
       title: company,
     );
   }
+
+  @override
+  Size get preferredSize => Size.fromHeight(50.h);
 }
