@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:job_hub/features/onboarding/presentation/views/onboarding_view/widgets/onboarding_view_body.dart';
 
 class OnboardingView extends StatelessWidget {
@@ -6,10 +7,16 @@ class OnboardingView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const SafeArea(
-      top: false,
-      child: Scaffold(
-        body: OnboardingViewBody(),
+    return PopScope(
+      canPop: true,
+      onPopInvoked: (value) {
+        SystemNavigator.pop();
+      },
+      child: const SafeArea(
+        top: false,
+        child: Scaffold(
+          body: OnboardingViewBody(),
+        ),
       ),
     );
   }
