@@ -14,8 +14,14 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'features/chat/data/repos/chat_repo/chat_repo_impl.dart';
 import 'features/home/data/repos/jobs_repo/jobs_repo_impl.dart';
 import 'features/chat/presentation/cubits/chat_cubit/chat_cubit.dart';
+import 'features/profile/data/repos/profile_repo/profile_repo_impl.dart';
+import 'features/drawer/presentation/cubits/drawer_cubit/drawer_cubit.dart';
+import 'features/profile/presentation/cubits/profile_cubit/profile_cubit.dart';
 import 'package:job_hub/features/bookmarks/data/repos/bookmarks_repo_impl.dart';
+import 'package:job_hub/features/auth/data/repos/auth_repo/auth_repo_impl.dart';
 import 'package:job_hub/features/bookmarks/presentation/cubits/bookmarks_cubit.dart';
+import 'package:job_hub/features/auth/presentation/cubits/media_cubit/media_cubit.dart';
+import 'package:job_hub/features/auth/presentation/cubits/register_cubit/register_cubit.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -27,23 +33,32 @@ void main() async {
   );
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
   await CacheHelper.init();
-  await handleCache();
+  await getCache();
+  await ScreenUtil.ensureScreenSize();
   setupServiceLocator();
   Bloc.observer = MyBlocObserver();
   runApp(
     MultiBlocProvider(
       providers: [
+        BlocProvider(create: (context) => CustomDrawerCubit()),
+        BlocProvider(create: (context) => MediaCubit()),
+        BlocProvider(
+          create: (context) => RegisterCubit(getIt.get<AuthRepoImpl>()),
+        ),
         BlocProvider(
           create: (context) =>
               JobsCubit(getIt.get<JobsRepoImpl>())..getAllJobs(),
         ),
         BlocProvider(
-          create: (context) =>
-              BookmarksCubit(getIt.get<BookmarksRepoImpl>())..getAllBookmarks(),
+          create: (context) => BookmarksCubit(getIt.get<BookmarksRepoImpl>()),
         ),
         BlocProvider(
           create: (context) =>
               ChatCubit(getIt.get<ChatRepoImpl>())..getAllChats(),
+        ),
+        BlocProvider(
+          create: (context) =>
+              ProfileCubit(getIt.get<ProfileRepoImpl>())..getUser(),
         ),
       ],
       child: const JobHub(),
@@ -60,6 +75,7 @@ class JobHub extends StatelessWidget {
       useInheritedMediaQuery: true,
       designSize: const Size(375, 812),
       minTextAdapt: true,
+      ensureScreenSize: true,
       splitScreenMode: true,
       builder: (context, child) {
         return MaterialApp.router(
